requirements.txt
flask[async]==2.3.2
aiohttp==3.9.3
youtube-transcript-api==0.6.1
hypercorn==0.16.0
python-dotenv==1.0.1
